% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plottableR.R
\name{plottable}
\alias{plottable}
\title{Plot R data with Plottable.js}
\usage{
plottable(data = NULL, components = NULL, scales = NULL, axes = NULL,
  plots = NULL, ..., width = NULL, height = NULL)
}
\description{
plotting and creation function.
}

